refactor(perl-lexer): split lexer helpers (#0000)
