+++ crates/perl-lexer/src/lexer/helpers/mod.rs
@@ -1,3 +1,5 @@
+mod balanced_segments;
+mod cursor;
 mod file_normalization;
 mod heredoc_delimiter;
 mod line_scanning;
