+++ crates/perl-lexer/src/lib.rs
@@ -591,208 +591,7 @@ impl<'a> PerlLexer<'a> {
         self.mode = LexerMode::InFormatBody;
     }
 
-    // Internal helper methods
-
-    #[allow(clippy::inline_always)] // Performance critical in lexer hot path
-    #[inline(always)]
-    fn byte_at(bytes: &[u8], index: usize) -> u8 {
-        debug_assert!(index < bytes.len());
-        match bytes.get(index) {
-            Some(&byte) => byte,
-            None => 0,
-        }
-    }
-
-    /// Ensure the internal byte offset points at a UTF-8 char boundary.
-    ///
-    /// This is a defensive guard against malformed intermediate offsets from
-    /// complex lookahead/backtracking paths so downstream slicing never panics.
-    #[inline]
-    fn normalize_char_boundary(&mut self) {
-        while self.position < self.input.len() && !self.input.is_char_boundary(self.position) {
-            self.position += 1;
-        }
-    }
-
-    #[allow(clippy::inline_always)] // Performance critical in lexer hot path
-    #[inline(always)]
-    fn current_char(&self) -> Option<char> {
-        if self.position < self.input_bytes.len() {
-            if !self.input.is_char_boundary(self.position) {
-                return None;
-            }
-            // For ASCII, direct access is safe
-            let byte = Self::byte_at(self.input_bytes, self.position);
-            if byte < 128 {
-                Some(byte as char)
-            } else {
-                // For non-ASCII, fall back to proper UTF-8 parsing
-                self.input.get(self.position..).and_then(|s| s.chars().next())
-            }
-        } else {
-            None
-        }
-    }
-
-    #[inline(always)]
-    fn peek_char(&self, offset: usize) -> Option<char> {
-        if offset > self.config.max_lookahead {
-            return None;
-        }
-        if !self.input.is_char_boundary(self.position) {
-            return None;
-        }
-
-        let pos = self.position.checked_add(offset)?;
-        if pos < self.input_bytes.len() {
-            // For ASCII, direct access is safe
-            let byte = Self::byte_at(self.input_bytes, pos);
-            if byte < 128 {
-                Some(byte as char)
-            } else {
-                // For non-ASCII, use chars iterator
-                self.input.get(self.position..).and_then(|s| s.chars().nth(offset))
-            }
-        } else {
-            None
-        }
-    }
-
-    #[allow(clippy::inline_always)] // Performance critical in lexer hot path
-    #[inline(always)]
-    fn advance(&mut self) {
-        if self.position < self.input_bytes.len() {
-            if !self.input.is_char_boundary(self.position) {
-                self.normalize_char_boundary();
-                return;
-            }
-            let byte = Self::byte_at(self.input_bytes, self.position);
-            if byte < 128 {
-                // ASCII fast path
-                self.position += 1;
-            } else if let Some(ch) = self.input.get(self.position..).and_then(|s| s.chars().next())
-            {
-                self.position += ch.len_utf8();
-            }
-        }
-    }
-
-    /// General-purpose balanced-segment consumer (no quote-boundary recovery).
-    ///
-    /// For use inside double-quoted string interpolation where the outer `"` must
-    /// act as a recovery boundary, use [`consume_balanced_segment_in_string`] instead.
-    #[allow(dead_code)]
-    #[inline]
-    fn consume_balanced_segment(&mut self, open: char, close: char) -> Option<usize> {
-        if self.current_char() != Some(open) {
-            return None;
-        }
-
-        let mut depth = 1usize;
-        self.advance();
-        while let Some(ch) = self.current_char() {
-            match ch {
-                '\\' => {
-                    self.advance();
-                    if self.current_char().is_some() {
-                        self.advance();
-                    }
-                }
-                c if c == open => {
-                    depth += 1;
-                    self.advance();
-                }
-                c if c == close => {
-                    self.advance();
-                    depth -= 1;
-                    if depth == 0 {
-                        return Some(self.position);
-                    }
-                }
-                _ => self.advance(),
-            }
-        }
-
-        None
-    }
-
-    #[inline]
-    fn consume_balanced_segment_in_string(
-        &mut self,
-        open: char,
-        close: char,
-        terminator: char,
-    ) -> Option<usize> {
-        if self.current_char() != Some(open) {
-            return None;
-        }
-
-        let mut depth = 1usize;
-        self.advance();
-        while let Some(ch) = self.current_char() {
-            match ch {
-                '\\' => {
-                    self.advance();
-                    if self.current_char().is_some() {
-                        self.advance();
-                    }
-                }
-                c if c == terminator => {
-                    // Local recovery for interpolation tails in quoted strings:
-                    // stop at the closing quote so the outer string parser can
-                    // still terminate this token cleanly.
-                    return None;
-                }
-                c if c == open => {
-                    depth += 1;
-                    self.advance();
-                }
-                c if c == close => {
-                    self.advance();
-                    depth -= 1;
-                    if depth == 0 {
-                        return Some(self.position);
-                    }
-                }
-                _ => self.advance(),
-            }
-        }
-
-        None
-    }
-
-    /// Fast byte-level check for ASCII characters
-    #[inline]
-    fn peek_byte(&self, offset: usize) -> Option<u8> {
-        if offset > self.config.max_lookahead {
-            return None;
-        }
-
-        let pos = self.position.checked_add(offset)?;
-        if pos < self.input_bytes.len() { Some(self.input_bytes[pos]) } else { None }
-    }
-
-    /// Check if the next bytes match a pattern (ASCII only)
-    #[inline]
-    fn matches_bytes(&self, pattern: &[u8]) -> bool {
-        let Some(end_offset) = pattern.len().checked_sub(1) else {
-            return true;
-        };
-
-        if end_offset > self.config.max_lookahead {
-            return false;
-        }
-
-        let Some(end) = self.position.checked_add(pattern.len()) else {
-            return false;
-        };
-
-        if end <= self.input_bytes.len() {
-            &self.input_bytes[self.position..end] == pattern
-        } else {
-            false
-        }
-    }
+    // Token-specific parsing methods
 
     #[inline]
     fn skip_whitespace_and_comments(&mut self) -> Option<()> {
